+++ nifti2bids/_decorators.py
@@ -2,9 +2,10 @@
 
 import functools, inspect
 
-from typing import Any, Callable
+from typing import Any, Callable, Optional
 
 from ._helpers import list_to_str
+from .io import get_nifti_header
 
 
 def check_all_none(parameter_names: list[str]) -> Callable:
@@ -45,6 +46,34 @@ def check_all_none(parameter_names: list[str]) -> Callable:
                     "All of the following arguments cannot be None, "
                     f"one must be specified: {list_to_str(parameter_names)}."
                 )
+
+            return func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
+
+
+def check_nifti(nifti_param_name: Optional[str] = None) -> Callable:
+    def decorator(func: Callable) -> Callable:
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs) -> Any:
+            bound_args = inspect.signature(func).bind(*args, **kwargs)
+            bound_args.apply_defaults()
+
+            param_name = (
+                "nifti_file_or_img" if not nifti_param_name else nifti_param_name
+            )
+            img_val = bound_args.arguments.get(param_name)
+            if img_val:
+                hdr = get_nifti_header(img_val)
+
+                # sform takes precedence over qform
+                code = "sform_code" if hdr["sform_code"] else "qform_code"
+                if hdr[code] != 1:
+                    raise ValueError(
+                        f"The {code} is not set to 'scanner' and is not a raw NIfTI image."
+                    )
 
             return func(*args, **kwargs)